Added storing orderDetails and bugfix for POST order

+++ order-service/src/main/java/com/bits/pilani/orderservice/dto/MenuItem.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @Embeddable
 public class MenuItem {
-    private Long id;
+    private Integer id;
     private String name;
-    private int quantity;
+    private Integer quantity;
 }
